Add EnvironmentTest
